requirements.txt
pytest==3.0.6
alembic==1.7.1
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
protobuf==3.17.3
pycparser==2.20
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0